[#6] - refactoring and phpcs
