+++ tests/cli/Cli/Router/Route/GetSetDescriptionCest.php
@@ -32,7 +32,9 @@ class GetSetDescriptionCest
         $router = new Router(false);
 
         foreach ($routes as $pattern => $description) {
-            $router->add($pattern)->setDescription($description);
+            $router->add($pattern)
+                   ->setDescription($description)
+            ;
         }
 
         foreach ($router->getRoutes() as $route) {
