+++ tests/cli/Di/FactoryDefault/Cli/ConstructCest.php
@@ -15,21 +15,13 @@ namespace Phalcon\Tests\Cli\Di\FactoryDefault\Cli;
 
 use CliTester;
 use Codeception\Example;
-use Phalcon\Annotations\Adapter\Memory as AnnotationsMemory;
-use Phalcon\Cli\Dispatcher;
-use Phalcon\Cli\Router;
 use Phalcon\Di\FactoryDefault\Cli;
-use Phalcon\Html\Escaper;
-use Phalcon\Events\Manager as EventsManager;
-use Phalcon\Filter\Filter;
-use Phalcon\Mvc\Model\Manager as ModelsManager;
-use Phalcon\Mvc\Model\MetaData\Memory;
-use Phalcon\Mvc\Model\Transaction\Manager;
-use Phalcon\Encryption\Security;
-use Phalcon\Support\HelperFactory;
+use Phalcon\Tests\Fixtures\Traits\CliTrait;
 
 class ConstructCest
 {
+    use CliTrait;
+
     /**
      * Tests Phalcon\Di\FactoryDefault\Cli :: __construct()
      *
@@ -71,55 +63,5 @@ class ConstructCest
         $class  = $example['class'];
         $actual = $container->get($example['service'], $params);
         $I->assertInstanceOf($class, $actual);
-    }
-
-    private function getServices(): array
-    {
-        return [
-//            [
-//                'service' => 'annotations',
-//                'class'   => AnnotationsMemory::class,
-//            ],
-            [
-                'service' => 'dispatcher',
-                'class'   => Dispatcher::class,
-            ],
-            [
-                'service' => 'escaper',
-                'class'   => Escaper::class,
-            ],
-            [
-                'service' => 'eventsManager',
-                'class'   => EventsManager::class,
-            ],
-            [
-                'service' => 'filter',
-                'class'   => Filter::class,
-            ],
-            [
-                'service' => 'helper',
-                'class'   => HelperFactory::class,
-            ],
-//            [
-//                'service' => 'modelsManager',
-//                'class'   => ModelsManager::class,
-//            ],
-//            [
-//                'service' => 'modelsMetadata',
-//                'class'   => Memory::class,
-//            ],
-            [
-                'service' => 'router',
-                'class'   => Router::class,
-            ],
-            [
-                'service' => 'security',
-                'class'   => Security::class,
-            ],
-//            [
-//                'service' => 'transactionManager',
-//                'class'   => Manager::class,
-//            ],
-        ];
     }
 }
